backend/pyproject.toml
[project]
name = "mobile-mapping-viewer-backend"
version = "0.1.0"
description = "MCP-first mobile mapping viewer backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.5",
    "uvicorn[standard]>=0.32.1",
    "pydantic>=2.10.3",
    "groq>=0.11.0",
    "python-dotenv>=1.0.1",
    "rio-tiler>=7.2.0",
    "rasterio>=1.4.3",
    "pillow>=11.0.0",
    "geopandas>=1.1.1",
    "fiona>=1.10.1",
]

